+++ tests/test_schema.py
@@ -1,6 +1,6 @@
 from pytest import mark
 
-from harp.schema import read_schema
+from harp.data.schema import read_schema
 from tests.params import DeviceSchemaParam
 
 testdata = [
